+++ examples/advanced/src/gen/clients/axios/petService/findPetsByStatus.ts
@@ -3,7 +3,8 @@ import fetch from '../../../../axios-client.ts'
 import type { FindPetsByStatus400, FindPetsByStatusPathParams, FindPetsByStatusQueryResponse } from '../../../models/ts/petController/FindPetsByStatus.ts'
 import { findPetsByStatusQueryResponseSchema } from '../../../zod/petController/findPetsByStatusSchema.ts'
 
-export function getFindPetsByStatusUrl({ step_id }: { step_id: FindPetsByStatusPathParams['step_id'] }) {
+export function getFindPetsByStatusUrl({ stepId }: { stepId: FindPetsByStatusPathParams['stepId'] }) {
+  const step_id = stepId
   const res = { method: 'GET', url: `https://petstore3.swagger.io/api/v3/pet/findByStatus/${step_id}` as const }
   return res
 }
@@ -14,14 +15,16 @@ export function getFindPetsByStatusUrl({ step_id }: { step_id: FindPetsByStatusP
  * {@link /pet/findByStatus/:step_id}
  */
 export async function findPetsByStatus(
-  { step_id }: { step_id: FindPetsByStatusPathParams['step_id'] },
+  { stepId }: { stepId: FindPetsByStatusPathParams['stepId'] },
   config: Partial<RequestConfig> & { client?: typeof fetch } = {},
 ) {
   const { client: request = fetch, ...requestConfig } = config
 
+  const _step_id = stepId
+
   const res = await request<FindPetsByStatusQueryResponse, ResponseErrorConfig<FindPetsByStatus400>, unknown>({
     method: 'GET',
-    url: getFindPetsByStatusUrl({ step_id }).url.toString(),
+    url: getFindPetsByStatusUrl({ stepId }).url.toString(),
     ...requestConfig,
   })
   return { ...res, data: findPetsByStatusQueryResponseSchema.parse(res.data) }