+++ examples/advanced/src/gen/mcp/petRequests/findPetsByStatus.ts
@@ -8,10 +8,12 @@ import type { FindPetsByStatus400, FindPetsByStatusPathParams, FindPetsByStatusQ
  * @summary Finds Pets by status
  * {@link /pet/findByStatus/:step_id}
  */
-export async function findPetsByStatusHandler({ stepId }: { stepId: FindPetsByStatusPathParams['step_id'] }): Promise<Promise<CallToolResult>> {
+export async function findPetsByStatusHandler({ stepId }: { stepId: FindPetsByStatusPathParams['stepId'] }): Promise<Promise<CallToolResult>> {
+  const step_id = stepId
+
   const res = await fetch<FindPetsByStatusQueryResponse, ResponseErrorConfig<FindPetsByStatus400>, unknown>({
     method: 'GET',
-    url: `/pet/findByStatus/${stepId}`,
+    url: `/pet/findByStatus/${step_id}`,
     baseURL: 'https://petstore.swagger.io/v2',
   })
   return {