+++ examples/advanced/src/gen/mcp/petsRequests/createPets.ts
@@ -24,15 +24,19 @@ export async function createPetsHandler({
   headers: CreatePetsHeaderParams
   params?: CreatePetsQueryParams
 }): Promise<Promise<CallToolResult>> {
+  const mappedParams = params ? { bool_param: params.boolParam } : undefined
+
+  const mappedHeaders = headers ? { 'X-EXAMPLE': headers.xEXAMPLE } : undefined
+
   const requestData = data
 
   const res = await fetch<CreatePetsMutationResponse, ResponseErrorConfig<Error>, CreatePetsMutationRequest>({
     method: 'POST',
     url: `/pets/${uuid}`,
     baseURL: 'https://petstore.swagger.io/v2',
-    params,
+    params: mappedParams,
     data: requestData,
-    headers: { ...headers },
+    headers: { ...mappedHeaders },
   })
   return {
     content: [