+++ packages/plugin-oas/src/utils/index.ts
@@ -3,6 +3,13 @@ export { getBanner } from './getBanner.ts'
 export { getComments } from './getComments.ts'
 export { getFooter } from './getFooter.ts'
 export { getImports } from './getImports.ts'
-export { getPathParams } from './getParams.ts'
+export {
+  getParamsMapping,
+  getPathParams,
+} from './getParams.ts'
 export { getSchemaFactory } from './getSchemaFactory.ts'
 export { getSchemas } from './getSchemas.ts'
+export {
+  applyParamsCasing,
+  isParameterSchema,
+} from './paramsCasing.ts'