+++ packages/plugin-react-query/src/plugin.ts
@@ -67,6 +67,7 @@ export const pluginReactQuery = definePlugin<PluginReactQuery>((options) => {
         dataReturnType: client?.dataReturnType ?? 'data',
         pathParamsType,
         importPath: clientImportPath,
+        paramsCasing,
       },
       infinite: infinite
         ? {